src/sample_ddms/requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
chardet==3.0.4
distlib==0.3.3
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
django-environ==0.8.1
filelock==3.3.2
idna==2.10
jsonschema==3.2.0
Markdown==3.3.4
pipenv==2021.5.29
platformdirs==2.4.0
pyrsistent==0.18.0
pytz==2021.3
PyYAML==5.3.1
requests==2.25.0
six==1.16.0
sqlparse==0.4.2
strict-rfc3339==0.7
tenacity==6.2.0
toposort==1.6
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
